Moved some sharable code for the faction leaders into Shared/Core
